% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Grouping.R
\name{Add_Grouping}
\alias{Add_Grouping}
\title{Add in grouping information to column names}
\usage{
Add_Grouping(data)
}
\arguments{
\item{data}{Dataframe to change}
}
\value{
Dataframe with new columns
}
\description{
Extra column naming function in case the original names are very long
}
\examples{
\dontrun{
Add_Grouping(df)
}
}
