% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boxplot_Proteins.R
\name{Boxplot_Proteins}
\alias{Boxplot_Proteins}
\title{Function to produce boxplots for proteins of interest}
\usage{
Boxplot_Proteins(
  data,
  DA_result,
  matrix,
  proteins,
  group1,
  group2,
  p_type = "raw"
)
}
\arguments{
\item{data}{The normalised data}

\item{DA_result}{Results from DA analysis}

\item{matrix}{Metadata}

\item{proteins}{vector of proteins to produce boxplots}

\item{group1}{First group in metadata (control)}

\item{group2}{Second group in metadata (treated)}

\item{p_type}{The statistic to show ("raw" or "adj")}
}
\value{
Boxplot and the long data used to make the boxplot for more customisation
}
\description{
Function to produce boxplots for proteins of interest
}
\examples{
\dontrun{
Boxplot_Proteins(df, DA_result, matrix, c("AQP4", "ACTN3", "LRP1", "PRKA2"), "control", "treated", "adj")
}
}
