% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Data.R
\name{Check_Data}
\alias{Check_Data}
\title{Function to review missingness of data}
\usage{
Check_Data(
  data,
  group_patterns = c("young", "old"),
  group_labels = NULL,
  fill_colors = c("#F08080", "#48D1CC"),
  summary_colors = c("#1E90FF", "#DC143C", "#8A2BE2")
)
}
\arguments{
\item{data}{The dataframe}

\item{group_patterns}{The group within your experiment}

\item{group_labels}{The labels for the figures for each groups}

\item{fill_colors}{Colours for plot A}

\item{summary_colors}{Colours for plot B}
}
\value{
A cowplot with two figures. A) Missing proteins per sample. B) Summary of missingness of data
}
\description{
Summarising the amount of missing data within a proteomics dataset
}
\examples{
\dontrun{
Check_Data(data = df,
           group_patterns = c("young", "old"))
}
}
