% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Filtering.R
\name{Check_Filtering}
\alias{Check_Filtering}
\title{Function to see the effects of filtering threshold}
\usage{
Check_Filtering(data, matrix, n_samples)
}
\arguments{
\item{data}{The non-filtered data}

\item{matrix}{The meta data}

\item{n_samples}{The number of samples per group}
}
\value{
A plot showing filtering effect
}
\description{
Function to see the effects of filtering threshold
}
\examples{
\dontrun{
Check_Filtering(df, matrix, 10)
}
}
