% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Differential_Abundance_Analysis.R
\name{Differential_Abundance_Analysis}
\alias{Differential_Abundance_Analysis}
\title{Function to perform differential abundance analysis using the limma package}
\usage{
Differential_Abundance_Analysis(
  normalised_data,
  sample_matrix,
  contrast_formula = NULL,
  adjust.method = "BH"
)
}
\arguments{
\item{normalised_data}{Normalised dataframe}

\item{sample_matrix}{Metadata of experiment}

\item{contrast_formula}{A charachter to specify how to do the DAA eg. ("disease - control")}

\item{adjust.method}{Method to adjust p-values ("None", "BH", "BY", "holm", "hochberg", "hommel", "bonferroni")}
}
\value{
Differential expression table
}
\description{
Function to perform differential abundance analysis using the limma package
}
\examples{
\dontrun{
Differential_Abundance_Analysis(normalised_df,
                                matrix,
                                "old - young",
                                adjust.method = "BH")
}

}
