% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dummy_Data.R
\name{Dummy_Data}
\alias{Dummy_Data}
\title{Generate a dummy dataset}
\usage{
Dummy_Data(
  n_proteins = 500,
  group1 = "Control",
  group2 = "Treated",
  n_replicates = 5,
  missing_frac = 0.2
)
}
\arguments{
\item{n_proteins}{Number of proteins}

\item{group1}{Name of group 1}

\item{group2}{Name of group 2}

\item{n_replicates}{Number of desired replicates}

\item{missing_frac}{Percentage of missing values}
}
\value{
A dummy dataset
}
\description{
Generate a dummy dataset
}
\examples{
\dontrun{
  Dummy_Data <- (n_proteins = 500,group1 = "Control",group2 = "Treated",n_replicates = 5,missing_frac = 0.2
}

}
