% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_Proteins.R
\name{Filter_Proteins}
\alias{Filter_Proteins}
\title{Filtering protein function}
\usage{
Filter_Proteins(data, matrix, min_samples)
}
\arguments{
\item{data}{The dataset to filter}

\item{matrix}{The metadata matrix}

\item{min_samples}{The minimum number of samples the protein needs to be in.}
}
\value{
A filtered dataset
}
\description{
Function to specify how you wish to filter out proteins.
}
\examples{
\dontrun{
  filtered_df <- Filter_Proteins(df, matrix, min_samples = 3)
}

}
