% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_To_Protein.R
\name{Gene_To_Protein}
\alias{Gene_To_Protein}
\title{Function to convert gene names into protein names}
\usage{
Gene_To_Protein(DA_result, species = "hsapiens")
}
\arguments{
\item{DA_result}{The result from DA analysis}

\item{species}{The species to look in. First letter of the genus then the species}
}
\value{
A new DA data with protein description included
}
\description{
Function to convert gene names into protein names
}
\examples{
\dontrun{
DA_results <- Differential_Abundance_Analysis(normalised_df,
                                              matrix,
                                             "old - young",
                                              adjust.method = "BH")
Updated_DA <- Gene_To_Protein(DA_results, "hsapiens")

}
}
