% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_from_DA.R
\name{GO_from_DA}
\alias{GO_from_DA}
\title{Function to input the results of differential expression into ClusterProfiler.}
\usage{
GO_from_DA(
  DA_summary,
  ontology = "BP",
  use_adj_p = TRUE,
  group_up = "Up",
  group_down = "Down",
  top_n = 10
)
}
\arguments{
\item{DA_summary}{The result of SummariseDA}

\item{ontology}{The type of enrichment analysis MF, CC or BP}

\item{use_adj_p}{Use proteins significant at p_adj (True) or raw_p (FALSE)}

\item{group_up}{The name of the group associated with up-regulation}

\item{group_down}{The name of the group associated with down-regulation}

\item{top_n}{The number of terms to show on the dotplot}
}
\value{
A dotplot from clusterprofiler
}
\description{
Use SummariseDA function before this
}
\examples{
\dontrun{
GO_from_DA(DA_summary, "BP", TRUE, "Old", "Young", 10)
}
}
