% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_from_DA_all.R
\name{GO_from_DA_all}
\alias{GO_from_DA_all}
\title{Function to run all the Gene Ontology analysis at once}
\usage{
GO_from_DA_all(group_up, group_down, DA_summary, save_file, name = NULL)
}
\arguments{
\item{group_up}{The group that corresponds to a positive fold change}

\item{group_down}{The group that corresponds to a negative fold change}

\item{DA_summary}{The DA summary object produced using the AndersR package}

\item{save_file}{The location to save images}

\item{name}{Name to help with saving files}
}
\value{
Multiple ggplots from clusterProfiler
}
\description{
Function to run all the Gene Ontology analysis at once
}
\examples{
\dontrun{
GO_from_DA_all("Old", "Young", DA_summary, "C:/file_path/plots/", "Experiment_1")
}
}
