% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_IPA.R
\name{Make_IPA}
\alias{Make_IPA}
\title{Function to make an IPA suitable file}
\usage{
Make_IPA(de_file, tsv_file)
}
\arguments{
\item{de_file}{Differential expression data generated with the AndersR package}

\item{tsv_file}{The TSV file containing all data (Make sure the gene names are the rownames)}
}
\value{
An IPA suitable dataset
}
\description{
Function to make an IPA suitable file
}
\examples{
\dontrun{
IPA_data <- Make_IPA("file_path_de.csv", "file_path_all_data.tsv")
}
}
