% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Matrix.R
\name{Make_Matrix}
\alias{Make_Matrix}
\title{Make a sample matrix}
\usage{
Make_Matrix(data)
}
\arguments{
\item{data}{Dataset with sample information}
}
\value{
A sample matrix for downstream analysis
}
\description{
Make a sample matrix
}
\examples{
\dontrun{
Make_Matrix(data = df)
}
}
