% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalise_Data.R
\name{Normalise_Data}
\alias{Normalise_Data}
\title{Normalisation function}
\usage{
Normalise_Data(filtered_df, method)
}
\arguments{
\item{filtered_df}{A filtered dataframe}

\item{method}{The method of normalisation (Median, Log2, Quantile or VSN)}
}
\value{
A normalised dataframe
}
\description{
This function utilises the NormalyserDE package to help with normalisation
}
\examples{
\dontrun{
Normalise_Data(filtered_df, method = "Median")
}
}
