% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_PCA.R
\name{Plot_PCA}
\alias{Plot_PCA}
\title{Plot a PCA from a normalised proteomics dataset}
\usage{
Plot_PCA(normalised_df, matrix, labels)
}
\arguments{
\item{normalised_df}{A normalised data frame generated with AndersR}

\item{matrix}{Sample matrix generated with AndersR}

\item{labels}{A vector of labels for your samples}
}
\value{
A ggplot
}
\description{
Simple function to plot the results from a principal component analysis
}
\examples{
\dontrun{
Plot_PCA(normalised_df, matrix, c("Old", "Young"))
}
}
