% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Protein_Assay.R
\name{Protein_Assay}
\alias{Protein_Assay}
\title{Reads and process protein assay from xlsx sheet}
\usage{
Protein_Assay(
  file_path,
  n_standard = NULL,
  standard_cols = NULL,
  standard_conc = NULL,
  n_samples = NULL,
  sample_order = NULL,
  dodgy_standard = NULL
)
}
\arguments{
\item{file_path}{Where the file is saved}

\item{n_standard}{How many replicates of the standards}

\item{standard_cols}{The number the columns are in}

\item{standard_conc}{Vector with standard concentrations}

\item{n_samples}{How many samples in the 96-well plate}

\item{sample_order}{A vector of sample order. This assumes you went across in triplicates.}

\item{dodgy_standard}{Added this in cases when one of your standards is a bit rubbish}
}
\value{
A list with all the analysis outputs
}
\description{
XLSX file from spectrostar Nano always produces the same sheet and this
function attempts to analyse it quickly based of my processing protocol.
}
\details{
This protocol hasn't let me down so far.
}
\examples{
Protein_Assay(file_path = "protein_assay_result.xlsx",
n_standard = 8,
standard_cols = c(1, 2, 3),
standard_conc = c(2000, 1500, 1000, 750, 500, 250, 125, 62.5),
n_samples = 24,
sample_order = LETTERS[1:24])

}
