% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarise_DA.R
\name{Summarise_DA}
\alias{Summarise_DA}
\title{Summarise DA analysis function}
\usage{
Summarise_DA(result, logFC_threshold = 1, p_value_threshold = 0.05)
}
\arguments{
\item{result}{The DA output from limma}

\item{logFC_threshold}{The logFC threshold (1)}

\item{p_value_threshold}{The pvalue threshold (0.05)}
}
\value{
A list with differentially abundant proteins
}
\description{
Takes the results of differential abundance analysis and extracts significant proteins.
}
\examples{
\dontrun{
Summary_result <- Summarise_DA(DA_results, 1, 0.05)
}
}
