% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_Normalisation.R
\name{Test_Normalisation}
\alias{Test_Normalisation}
\title{Normalisation comparison function}
\usage{
Test_Normalisation(filtered_df, sample_matrix)
}
\arguments{
\item{filtered_df}{Filtered dataframe}

\item{sample_matrix}{Metadata for your experiment}
}
\value{
Coefficient of variation, Density, Principal component analysis, MA and Relative log expression plots
}
\description{
Normalisation comparison function
}
\examples{
\dontrun{
Test_Normalisation(filtered_df, matrix)
}
}
