% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Volcano_Plot.R
\name{Volcano_Plot}
\alias{Volcano_Plot}
\title{Produce a volcano plot function}
\usage{
Volcano_Plot(daa_result, pval_threshold = 0.05, logFC_threshold = 1)
}
\arguments{
\item{daa_result}{Result from limma differential abundance analysis}

\item{pval_threshold}{threshold for the p-value (0.05)}

\item{logFC_threshold}{threshold for log2FC (1)}
}
\value{
A volcano plot
}
\description{
Produce a volcano plot function
}
\examples{
\dontrun{
  Volcano_Plot(result, 0.05, 1)
}
}
