% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_Proteins.R
\name{Filter_Proteins}
\alias{Filter_Proteins}
\title{Filtering protein function}
\usage{
Filter_Proteins(data, matrix, min_frac = 0.7)
}
\arguments{
\item{data}{The dataset to filter}

\item{matrix}{The metadata matrix}

\item{min_frac}{The percent of proteins to filter out (0 - 1)}
}
\value{
A filtered dataset
}
\description{
Function to specify how you wish to filter out proteins.
}
\examples{
\dontrun{
  filtered_df <- Filter_Proteins(df, matrix, min_frac = 0.5)
}

}
